refactor(node): split NodeStats helpers
Extract config, HTML rendering, store stats, and fieldset export into dedicated classes to reduce NodeStats dependencies and simplify usage.\n\nUpdate call sites and tests to use top-level load stat and request snapshot types.
