+++ src/main/java/network/crypta/node/NodeStatsFieldSetExporter.java
@@ -1,0 +1,388 @@
+package network.crypta.node;
+
+import static java.util.concurrent.TimeUnit.DAYS;
+
+import network.crypta.node.RequestTracker.WaitingForSlots;
+import network.crypta.support.SimpleFieldSet;
+
+/**
+ * Builds volatile statistics snapshots for node and peer reporting.
+ *
+ * <p>Extracted from {@link NodeStats} to keep the core collector focused on accumulation logic.
+ */
+final class NodeStatsFieldSetExporter {
+
+  private NodeStatsFieldSetExporter() {}
+
+  static SimpleFieldSet exportVolatileFieldSet(NodeStats stats) {
+    SimpleFieldSet fs = new SimpleFieldSet(true);
+    long now = System.currentTimeMillis();
+    fs.put("isUsingWrapper", stats.node.isUsingWrapper());
+    putUptime(stats, fs, now);
+    putPingAndDelays(stats, fs);
+    long nodeUptimeSecondsLocal = Math.max(1, (now - stats.node.getStartupTime()) / 1000);
+    putNetworkSizeEstimates(stats, fs, now);
+    putRoutingMissDistances(stats, fs);
+
+    fs.put("backedOffPercent", stats.backedOffPercent.currentValue());
+    fs.put("pInstantReject", stats.pRejectIncomingInstantly());
+    fs.put("unclaimedFIFOSize", stats.node.getUSM().getUnclaimedFIFOSize());
+    fs.put("RAMBucketPoolSize", stats.node.getClientCore().getTempBucketFactory().getRamUsed());
+
+    /* gather connection statistics */
+    PeerNodeStatus[] peerNodeStatuses = stats.peers.statusBook().getPeerNodeStatuses(true);
+    int numberOfSeedServers = 0;
+    int numberOfSeedClients = 0;
+
+    for (PeerNodeStatus peerNodeStatus : peerNodeStatuses) {
+      if (peerNodeStatus.isSeedServer()) numberOfSeedServers++;
+      if (peerNodeStatus.isSeedClient()) numberOfSeedClients++;
+    }
+
+    int numberOfConnected =
+        PeerNodeStatus.getPeerStatusCount(peerNodeStatuses, PeerManager.PEER_NODE_STATUS_CONNECTED);
+    int numberOfRoutingBackedOff =
+        PeerNodeStatus.getPeerStatusCount(
+            peerNodeStatuses, PeerManager.PEER_NODE_STATUS_ROUTING_BACKED_OFF);
+    int numberOfTooNew =
+        PeerNodeStatus.getPeerStatusCount(peerNodeStatuses, PeerManager.PEER_NODE_STATUS_TOO_NEW);
+    int numberOfTooOld =
+        PeerNodeStatus.getPeerStatusCount(peerNodeStatuses, PeerManager.PEER_NODE_STATUS_TOO_OLD);
+    int numberOfDisconnected =
+        PeerNodeStatus.getPeerStatusCount(
+            peerNodeStatuses, PeerManager.PEER_NODE_STATUS_DISCONNECTED);
+    int numberOfNeverConnected =
+        PeerNodeStatus.getPeerStatusCount(
+            peerNodeStatuses, PeerManager.PEER_NODE_STATUS_NEVER_CONNECTED);
+    int numberOfDisabled =
+        PeerNodeStatus.getPeerStatusCount(peerNodeStatuses, PeerManager.PEER_NODE_STATUS_DISABLED);
+    int numberOfBursting =
+        PeerNodeStatus.getPeerStatusCount(peerNodeStatuses, PeerManager.PEER_NODE_STATUS_BURSTING);
+    int numberOfListening =
+        PeerNodeStatus.getPeerStatusCount(peerNodeStatuses, PeerManager.PEER_NODE_STATUS_LISTENING);
+    int numberOfListenOnly =
+        PeerNodeStatus.getPeerStatusCount(
+            peerNodeStatuses, PeerManager.PEER_NODE_STATUS_LISTEN_ONLY);
+
+    int numberOfSimpleConnected = numberOfConnected + numberOfRoutingBackedOff;
+    int numberOfNotConnected =
+        numberOfTooNew
+            + numberOfTooOld
+            + numberOfDisconnected
+            + numberOfNeverConnected
+            + numberOfDisabled
+            + numberOfBursting
+            + numberOfListening
+            + numberOfListenOnly;
+
+    fs.put("numberOfSeedServers", numberOfSeedServers);
+    fs.put("numberOfSeedClients", numberOfSeedClients);
+    fs.put("numberOfConnected", numberOfConnected);
+    fs.put("numberOfRoutingBackedOff", numberOfRoutingBackedOff);
+    fs.put("numberOfTooNew", numberOfTooNew);
+    fs.put("numberOfTooOld", numberOfTooOld);
+    fs.put("numberOfDisconnected", numberOfDisconnected);
+    fs.put("numberOfNeverConnected", numberOfNeverConnected);
+    fs.put("numberOfDisabled", numberOfDisabled);
+    fs.put("numberOfBursting", numberOfBursting);
+    fs.put("numberOfListening", numberOfListening);
+    fs.put("numberOfListenOnly", numberOfListenOnly);
+
+    fs.put("numberOfSimpleConnected", numberOfSimpleConnected);
+    fs.put("numberOfNotConnected", numberOfNotConnected);
+
+    fs.put(
+        "numberOfTransferringRequestSenders",
+        stats.node.getTracker().getNumTransferringRequestSenders());
+    fs.put("numberOfARKFetchers", stats.node.getNumARKFetchers());
+    fs.put(
+        "bandwidthLiabilityUsageOutputBulk",
+        stats.node.getNodeStats().getBandwidthLiabilityUsage());
+
+    RequestTracker tracker = stats.node.getTracker();
+
+    fs.put("numberOfLocalCHKInserts", tracker.getNumLocalCHKInserts());
+    fs.put("numberOfRemoteCHKInserts", tracker.getNumRemoteCHKInserts());
+    fs.put("numberOfLocalSSKInserts", tracker.getNumLocalSSKInserts());
+    fs.put("numberOfRemoteSSKInserts", tracker.getNumRemoteSSKInserts());
+    fs.put("numberOfLocalCHKRequests", tracker.getNumLocalCHKRequests());
+    fs.put("numberOfRemoteCHKRequests", tracker.getNumRemoteCHKRequests());
+    fs.put("numberOfLocalSSKRequests", tracker.getNumLocalSSKRequests());
+    fs.put("numberOfRemoteSSKRequests", tracker.getNumRemoteSSKRequests());
+    fs.put(
+        "numberOfTransferringRequestHandlers",
+        stats.node.getTracker().getNumTransferringRequestHandlers());
+    fs.put("numberOfCHKOfferReplys", tracker.getNumCHKOfferReplies());
+    fs.put("numberOfSSKOfferReplys", tracker.getNumSSKOfferReplies());
+
+    double[] delayValues = stats.getNlmDelaySnapshot();
+    fs.put("delayTimeLocalRT", delayValues[0]);
+    fs.put("delayTimeRemoteRT", delayValues[1]);
+    fs.put("delayTimeLocalBulk", delayValues[2]);
+    fs.put("delayTimeRemoteBulk", delayValues[3]);
+    long[] slotTimeouts = stats.getSlotTimeoutSnapshot();
+    // timeoutFractions = fatalTimeouts/(fatalTimeouts+allocatedSlot)
+    fs.put("fatalTimeoutsLocal", slotTimeouts[0]);
+    fs.put("fatalTimeoutsRemote", slotTimeouts[1]);
+    fs.put("allocatedSlotLocal", slotTimeouts[2]);
+    fs.put("allocatedSlotRemote", slotTimeouts[3]);
+
+    WaitingForSlots waitingSlots = tracker.countRequestsWaitingForSlots();
+    fs.put("RequestsWaitingSlotsLocal", waitingSlots.local);
+    fs.put("RequestsWaitingSlotsRemote", waitingSlots.remote);
+
+    fs.put(
+        "successfulLocalCHKFetchTimeBulk",
+        stats.successfulLocalCHKFetchTimeAverageBulk.currentValue());
+    fs.put(
+        "successfulLocalCHKFetchTimeRT", stats.successfulLocalCHKFetchTimeAverageRT.currentValue());
+    fs.put(
+        "unsuccessfulLocalCHKFetchTimeBulk",
+        stats.unsuccessfulLocalCHKFetchTimeAverageBulk.currentValue());
+    fs.put(
+        "unsuccessfulLocalCHKFetchTimeRT",
+        stats.unsuccessfulLocalCHKFetchTimeAverageRT.currentValue());
+
+    fs.put(
+        "successfulLocalSSKFetchTimeBulk",
+        stats.successfulLocalSSKFetchTimeAverageBulk.currentValue());
+    fs.put(
+        "successfulLocalSSKFetchTimeRT", stats.successfulLocalSSKFetchTimeAverageRT.currentValue());
+    fs.put(
+        "unsuccessfulLocalSSKFetchTimeBulk",
+        stats.unsuccessfulLocalSSKFetchTimeAverageBulk.currentValue());
+    fs.put(
+        "unsuccessfulLocalSSKFetchTimeRT",
+        stats.unsuccessfulLocalSSKFetchTimeAverageRT.currentValue());
+
+    putTotalAndRecentIOMetrics(stats, fs, nodeUptimeSecondsLocal);
+
+    putRoutingBackoffCounters(stats, fs);
+
+    putSwapAndStoreMetrics(stats, fs, nodeUptimeSecondsLocal);
+
+    Runtime rt = Runtime.getRuntime();
+    float freeMemory = rt.freeMemory();
+    float totalMemory = rt.totalMemory();
+    float maxMemory = rt.maxMemory();
+
+    long usedJavaMem = (long) (totalMemory - freeMemory);
+    long allocatedJavaMem = (long) totalMemory;
+    long maxJavaMem = (long) maxMemory;
+    int availableCpus = rt.availableProcessors();
+
+    fs.put("freeJavaMemory", (long) freeMemory);
+    fs.put("usedJavaMemory", usedJavaMem);
+    fs.put("allocatedJavaMemory", allocatedJavaMem);
+    fs.put("maximumJavaMemory", maxJavaMem);
+    fs.put("availableCPUs", availableCpus);
+    fs.put("runningThreadCount", stats.getActiveThreadCount());
+
+    fs.put("globalFetchPSuccess", stats.globalFetchPSuccess.currentValue());
+    fs.put("globalFetchCount", stats.globalFetchPSuccess.countReports());
+    fs.put("chkLocalFetchPSuccess", stats.chkLocalFetchPSuccess.currentValue());
+    fs.put("chkLocalFetchCount", stats.chkLocalFetchPSuccess.countReports());
+    fs.put("chkRemoteFetchPSuccess", stats.chkRemoteFetchPSuccess.currentValue());
+    fs.put("chkRemoteFetchCount", stats.chkRemoteFetchPSuccess.countReports());
+    fs.put("sskLocalFetchPSuccess", stats.sskLocalFetchPSuccess.currentValue());
+    fs.put("sskLocalFetchCount", stats.sskLocalFetchPSuccess.countReports());
+    fs.put("sskRemoteFetchPSuccess", stats.sskRemoteFetchPSuccess.currentValue());
+    fs.put("sskRemoteFetchCount", stats.sskRemoteFetchPSuccess.countReports());
+    fs.put("blockTransferPSuccessRT", stats.blockTransferPSuccessRT.currentValue());
+    fs.put("blockTransferCountRT", stats.blockTransferPSuccessRT.countReports());
+    fs.put("blockTransferPSuccessBulk", stats.blockTransferPSuccessBulk.currentValue());
+    fs.put("blockTransferCountBulk", stats.blockTransferPSuccessBulk.countReports());
+    fs.put("blockTransferFailTimeout", stats.blockTransferFailTimeout.currentValue());
+
+    return fs;
+  }
+
+  private static void putUptime(NodeStats stats, SimpleFieldSet fs, long now) {
+    long nodeUptimeSeconds;
+    synchronized (stats) {
+      fs.put("startupTime", stats.node.getStartupTime());
+      nodeUptimeSeconds = (now - stats.node.getStartupTime()) / 1000;
+      if (nodeUptimeSeconds == 0) nodeUptimeSeconds = 1; // prevent division by zero
+      fs.put("uptimeSeconds", nodeUptimeSeconds);
+    }
+  }
+
+  private static void putPingAndDelays(NodeStats stats, SimpleFieldSet fs) {
+    fs.put("averagePingTime", stats.getNodeAveragePingTime());
+    fs.put("bwlimitDelayTime", stats.getBwlimitDelayTime());
+    fs.put("bwlimitDelayTimeRT", stats.getBwlimitDelayTimeRT());
+    fs.put("bwlimitDelayTimeBulk", stats.getBwlimitDelayTimeBulk());
+  }
+
+  private static void putNetworkSizeEstimates(NodeStats stats, SimpleFieldSet fs, long now) {
+    fs.put("opennetSizeEstimateSession", stats.getOpennetSizeEstimate(-1));
+    fs.put("networkSizeEstimateSession", stats.getDarknetSizeEstimate(-1));
+    for (int t = 1; t < 7; t++) {
+      int hour = t * 24;
+      long limit = now - DAYS.toMillis(t);
+      fs.put("opennetSizeEstimate" + hour + "hourRecent", stats.getOpennetSizeEstimate(limit));
+      fs.put("networkSizeEstimate" + hour + "hourRecent", stats.getDarknetSizeEstimate(limit));
+    }
+  }
+
+  private static void putRoutingMissDistances(NodeStats stats, SimpleFieldSet fs) {
+    fs.put("routingMissDistanceLocal", stats.routingMissDistanceLocal.currentValue());
+    fs.put("routingMissDistanceRemote", stats.routingMissDistanceRemote.currentValue());
+    fs.put("routingMissDistanceOverall", stats.routingMissDistanceOverall.currentValue());
+    fs.put("routingMissDistanceBulk", stats.routingMissDistanceBulk.currentValue());
+    fs.put("routingMissDistanceRT", stats.routingMissDistanceRT.currentValue());
+  }
+
+  private static void putTotalAndRecentIOMetrics(
+      NodeStats stats, SimpleFieldSet fs, long nodeUptimeSecondsLocal) {
+    long[] total = stats.node.getCollector().getTotalIO();
+    long totalOutputRate = (total[0]) / nodeUptimeSecondsLocal;
+    long totalInputRate = (total[1]) / nodeUptimeSecondsLocal;
+    long totalPayloadOutput = stats.node.getTotalPayloadSent();
+    long totalPayloadOutputRate = totalPayloadOutput / nodeUptimeSecondsLocal;
+    int totalPayloadOutputPercent =
+        (total[0] == 0) ? -1 : (int) (100 * totalPayloadOutput / total[0]);
+    fs.put("totalOutputBytes", total[0]);
+    fs.put("totalOutputRate", totalOutputRate);
+    fs.put("totalPayloadOutputBytes", totalPayloadOutput);
+    fs.put("totalPayloadOutputRate", totalPayloadOutputRate);
+    fs.put("totalPayloadOutputPercent", totalPayloadOutputPercent);
+    fs.put("totalInputBytes", total[1]);
+    fs.put("totalInputRate", totalInputRate);
+
+    long[] rate = stats.getNodeIOStats();
+    long deltaMS = (rate[5] - rate[2]);
+    double recentOutputRate = deltaMS == 0 ? 0 : (1000.0 * (rate[3] - rate[0]) / deltaMS);
+    double recentInputRate = deltaMS == 0 ? 0 : (1000.0 * (rate[4] - rate[1]) / deltaMS);
+    fs.put("recentOutputRate", recentOutputRate);
+    fs.put("recentInputRate", recentInputRate);
+
+    fs.put("ackOnlyBytes", stats.getNotificationOnlyPacketsSentBytes());
+    fs.put("resentBytes", stats.getResendBytesSent());
+    fs.put("updaterOutputBytes", stats.getUOMBytesSent());
+    fs.put("announcePayloadBytes", stats.getAnnounceBytesPayloadSent());
+    fs.put("announceSentBytes", stats.getAnnounceBytesSent());
+  }
+
+  private static void putRoutingBackoffCounters(NodeStats stats, SimpleFieldSet fs) {
+    String[] routingBackoffReasons = stats.peers.getPeerNodeRoutingBackoffReasons(true);
+    for (String routingBackoffReason : routingBackoffReasons) {
+      fs.put(
+          "numberWithRoutingBackoffReasonsRT." + routingBackoffReason,
+          stats.peers.getPeerNodeRoutingBackoffReasonSize(routingBackoffReason, true));
+    }
+
+    routingBackoffReasons = stats.peers.getPeerNodeRoutingBackoffReasons(false);
+    for (String routingBackoffReason : routingBackoffReasons) {
+      fs.put(
+          "numberWithRoutingBackoffReasonsBulk." + routingBackoffReason,
+          stats.peers.getPeerNodeRoutingBackoffReasonSize(routingBackoffReason, false));
+    }
+  }
+
+  private static void putSwapAndStoreMetrics(
+      NodeStats stats, SimpleFieldSet fs, long nodeUptimeSecondsLocal) {
+    double swaps = stats.node.getSwaps();
+    double noSwaps = stats.node.getNoSwaps();
+    double numberOfRemotePeerLocationsSeenInSwaps =
+        stats.node.getNumberOfRemotePeerLocationsSeenInSwaps();
+    fs.put("numberOfRemotePeerLocationsSeenInSwaps", numberOfRemotePeerLocationsSeenInSwaps);
+    double avgConnectedPeersPerNode = 0.0;
+    if ((numberOfRemotePeerLocationsSeenInSwaps > 0.0) && ((swaps > 0.0) || (noSwaps > 0.0))) {
+      avgConnectedPeersPerNode = numberOfRemotePeerLocationsSeenInSwaps / (swaps + noSwaps);
+    }
+    fs.put("avgConnectedPeersPerNode", avgConnectedPeersPerNode);
+
+    int startedSwaps = stats.node.getStartedSwaps();
+    int swapsRejectedAlreadyLocked = stats.node.getSwapsRejectedAlreadyLocked();
+    int swapsRejectedNowhereToGo = stats.node.getSwapsRejectedNowhereToGo();
+    int swapsRejectedRateLimit = stats.node.getSwapsRejectedRateLimit();
+    int swapsRejectedRecognizedID = stats.node.getSwapsRejectedRecognizedID();
+    double locationChangePerSession = stats.node.getLocationChangeSession();
+    double locationChangePerSwap = 0.0;
+    double locationChangePerMinute = 0.0;
+    double swapsPerMinute = 0.0;
+    double noSwapsPerMinute = 0.0;
+    double swapsPerNoSwaps = 0.0;
+    if (swaps > 0) {
+      locationChangePerSwap = locationChangePerSession / swaps;
+    }
+    if ((swaps > 0.0) && (nodeUptimeSecondsLocal >= 60)) {
+      locationChangePerMinute = locationChangePerSession / (nodeUptimeSecondsLocal / 60.0);
+    }
+    if ((swaps > 0.0) && (nodeUptimeSecondsLocal >= 60)) {
+      swapsPerMinute = swaps / (nodeUptimeSecondsLocal / 60.0);
+    }
+    if ((noSwaps > 0.0) && (nodeUptimeSecondsLocal >= 60)) {
+      noSwapsPerMinute = noSwaps / (nodeUptimeSecondsLocal / 60.0);
+    }
+    if ((swaps > 0.0) && (noSwaps > 0.0)) {
+      swapsPerNoSwaps = swaps / noSwaps;
+    }
+    fs.put("locationChangePerSession", locationChangePerSession);
+    fs.put("locationChangePerSwap", locationChangePerSwap);
+    fs.put("locationChangePerMinute", locationChangePerMinute);
+    fs.put("swapsPerMinute", swapsPerMinute);
+    fs.put("noSwapsPerMinute", noSwapsPerMinute);
+    fs.put("swapsPerNoSwaps", swapsPerNoSwaps);
+    fs.put("swaps", swaps);
+    fs.put("noSwaps", noSwaps);
+    fs.put("startedSwaps", startedSwaps);
+    fs.put("swapsRejectedAlreadyLocked", swapsRejectedAlreadyLocked);
+    fs.put("swapsRejectedNowhereToGo", swapsRejectedNowhereToGo);
+    fs.put("swapsRejectedRateLimit", swapsRejectedRateLimit);
+    fs.put("swapsRejectedRecognizedID", swapsRejectedRecognizedID);
+    long fix32kb = 32L * 1024;
+    long cachedKeys = stats.node.getChkDatacache().keyCount();
+    long cachedSize = cachedKeys * fix32kb;
+    long storeKeys = stats.node.getChkDatastore().keyCount();
+    long storeSize = storeKeys * fix32kb;
+    long overallKeys = cachedKeys + storeKeys;
+    long overallSize = cachedSize + storeSize;
+
+    long maxOverallKeys = stats.node.getMaxTotalKeys();
+    long maxOverallSize = maxOverallKeys * fix32kb;
+
+    double percentOverallKeysOfMax = (double) (overallKeys * 100) / (double) maxOverallKeys;
+
+    long cachedStoreHits = stats.node.getChkDatacache().hits();
+    long cachedStoreMisses = stats.node.getChkDatacache().misses();
+    long cachedStoreWrites = stats.node.getChkDatacache().writes();
+    long cacheAccesses = cachedStoreHits + cachedStoreMisses;
+    long cachedStoreFalsePositives = stats.node.getChkDatacache().getBloomFalsePositive();
+    double percentCachedStoreHitsOfAccesses =
+        (double) (cachedStoreHits * 100) / (double) cacheAccesses;
+    long storeHits = stats.node.getChkDatastore().hits();
+    long storeMisses = stats.node.getChkDatastore().misses();
+    long storeWrites = stats.node.getChkDatastore().writes();
+    long storeFalsePositives = stats.node.getChkDatastore().getBloomFalsePositive();
+    long storeAccesses = storeHits + storeMisses;
+    double percentStoreHitsOfAccesses = (double) (storeHits * 100) / (double) storeAccesses;
+    long overallAccesses = storeAccesses + cacheAccesses;
+    double avgStoreAccessRate = (double) overallAccesses / (double) nodeUptimeSecondsLocal;
+
+    fs.put("cachedKeys", cachedKeys);
+    fs.put("cachedSize", cachedSize);
+    fs.put("storeKeys", storeKeys);
+    fs.put("storeSize", storeSize);
+    fs.put("overallKeys", overallKeys);
+    fs.put("overallSize", overallSize);
+    fs.put("maxOverallKeys", maxOverallKeys);
+    fs.put("maxOverallSize", maxOverallSize);
+    fs.put("percentOverallKeysOfMax", percentOverallKeysOfMax);
+    fs.put("cachedStoreHits", cachedStoreHits);
+    fs.put("cachedStoreMisses", cachedStoreMisses);
+    fs.put("cachedStoreWrites", cachedStoreWrites);
+    fs.put("cacheAccesses", cacheAccesses);
+    fs.put("cachedStoreFalsePositives", cachedStoreFalsePositives);
+    fs.put("percentCachedStoreHitsOfAccesses", percentCachedStoreHitsOfAccesses);
+    fs.put("storeHits", storeHits);
+    fs.put("storeMisses", storeMisses);
+    fs.put("storeAccesses", storeAccesses);
+    fs.put("storeWrites", storeWrites);
+    fs.put("storeFalsePositives", storeFalsePositives);
+    fs.put("percentStoreHitsOfAccesses", percentStoreHitsOfAccesses);
+    fs.put("overallAccesses", overallAccesses);
+    fs.put("avgStoreAccessRate", avgStoreAccessRate);
+  }
+}
