+++ src/main/java/network/crypta/node/PeerNode.java
@@ -67,9 +67,7 @@ import network.crypta.keys.ClientSSK;
 import network.crypta.keys.FreenetURI;
 import network.crypta.keys.Key;
 import network.crypta.keys.USK;
-import network.crypta.node.NodeStats.PeerLoadStats;
 import network.crypta.node.NodeStats.RequestType;
-import network.crypta.node.NodeStats.RunningRequestsSnapshot;
 import network.crypta.node.OpennetManager.ConnectionType;
 import network.crypta.support.Base64;
 import network.crypta.support.BooleanLastTrueTracker;
